requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
httpx>=0.27.0
sqlalchemy>=2.0.30
aiosqlite>=0.20.0
slowapi>=0.1.9
structlog>=24.1.0
orjson>=3.10.0
numpy>=1.26.0
tenacity>=8.3.0
prometheus-client>=0.20.0
boto3>=1.35.0
google-genai>=1.0.0
